refactor: 대기 예약 순번을 JPA 쿼리에서 직접 계산하도록 변경
- 서비스 로직의 복잡도와 DB 접근 횟수를 줄임

+++ src/main/java/roomescape/controller/response/MemberReservationResponse.java
@@ -26,7 +26,7 @@ public record MemberReservationResponse(
 
     private static String formatStatus(WaitingWithRank waitingWithRank) {
         ReservationStatus status = waitingWithRank.reservationStatus();
-        int rank = waitingWithRank.rank();
+        long rank = waitingWithRank.rank();
 
         StringBuilder sb = new StringBuilder();
         if(status == ReservationStatus.WAITING) {
